refactor: support infinite scrolling and split code
